ICD11 - updates to data files

+++ mojo/src/main/java/gov/nih/nlm/ihtsdo/mapping/icd10cm/LookupIcd11MatchesMojo.java
@@ -62,7 +62,7 @@ public class LookupIcd11MatchesMojo extends AbstractMojo {
    * 
    * @parameter
    */
-  private long scoreThreshold;
+  private double scoreThreshold;
 
   /**
    * Executes the plugin.
